Moved localized elements from projectsConstants into their own constants file.
